pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leafar"
version = "0.1.0"
description = "Intelligent Android development CLI powered by Rafael Alves"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "mcp>=1.0.0",
]

[project.scripts]
rf = "leafar.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["leafar*"]
